requirements.txt
requests
pandas
Flask
gunicorn
google-cloud-storage
google-cloud-secret-manager
pytest
pytest-mock
python-dotenv